pyproject.toml
[project]
name = "madsci"
version = "0.6.2"
dependencies = ["pre-commit>=4.3.0", "structlog>=25.5.0", "colorama>=0.4.6"]
readme="README.md"
requires-python = ">=3.10.0"

[project.optional-dependencies]
docs = [
    "mystmd>=1.7.1",
    "pdoc3>=0.11.6",
]
[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock-resources[docker,mongo]>=2.12.2",
    "ipykernel>=6.29.5",
    "pre-commit>=4.3.0",
    "pydantic-settings-export>=1.0.2",
    "nbconvert>=7.16.6",
    "pytest-timeout>=2.4.0",
    "ruff>=0.14.9",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-asgi>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
timeout = 10

[tool.pydantic_settings_export]
project_dir = "."
default_settings = [
    "madsci.common.types.docker_types:DockerComposeSettings",
    "madsci.common.types.base_types:MadsciDeveloperSettings",
    "madsci.common.types.context_types:MadsciContext",
    "madsci.common.types.datapoint_types:DataManagerSettings",
    "madsci.common.types.datapoint_types:ObjectStorageSettings",
    "madsci.common.types.event_types:EventManagerSettings",
    "madsci.common.types.workcell_types:WorkcellManagerSettings",
    "madsci.common.types.experiment_types:ExperimentManagerSettings",
    "madsci.common.types.resource_types.definitions:ResourceManagerSettings",
    "madsci.common.types.lab_types:LabManagerSettings",
    "madsci.common.types.location_types:LocationManagerSettings",
    "madsci.common.types.event_types:EventClientConfig",
    "madsci.common.types.client_types:RestNodeClientConfig",
    "madsci.common.types.client_types:ResourceClientConfig",
    "madsci.common.types.client_types:ExperimentClientConfig",
    "madsci.common.types.client_types:DataClientConfig",
    "madsci.common.types.client_types:MadsciClientConfig",
    "madsci.common.types.client_types:LocationClientConfig",
    "madsci.common.types.client_types:WorkcellClientConfig",
    "madsci.common.types.client_types:LabClientConfig",
    "madsci.common.types.backup_types:PostgreSQLBackupSettings",
    "madsci.common.types.backup_types:MongoDBBackupSettings",
    "madsci.common.types.mongodb_migration_types:MongoDBMigrationSettings",
    "madsci.resource_manager.migration_tool:DatabaseMigrationSettings",
]

# Generate Markdown docs
[[tool.pydantic_settings_export.generators.markdown]]
paths = ["docs/Configuration.md"]

# Generate .env example
# Note: Default configuration should go in settings.yaml (version-controlled).
# The .env file is for secrets and environment-specific overrides only (gitignored).
# Manager settings support prefixed keys in YAML (e.g., event_server_url) via
# the alias_generator. Use `madsci config export` to generate settings.yaml content.
[[tool.pydantic_settings_export.generators.dotenv]]
paths = [".env.example"]

[tool.pdm.dev-dependencies]
dev = [
    "-e madsci.common @ file:///${PROJECT_ROOT}/src/madsci_common",
    "-e madsci.squid @ file:///${PROJECT_ROOT}/src/madsci_squid",
    "-e madsci.client @ file:///${PROJECT_ROOT}/src/madsci_client",
    "-e madsci.node_module @ file:///${PROJECT_ROOT}/src/madsci_node_module",
    "-e madsci.resource_manager @ file:///${PROJECT_ROOT}/src/madsci_resource_manager",
    "-e madsci.event_manager @ file:///${PROJECT_ROOT}/src/madsci_event_manager",
    "-e madsci.workcell_manager @ file:///${PROJECT_ROOT}/src/madsci_workcell_manager",
    "-e madsci.experiment_manager @ file:///${PROJECT_ROOT}/src/madsci_experiment_manager",
    "-e madsci.data_manager @ file:///${PROJECT_ROOT}/src/madsci_data_manager",
    "-e madsci.location_manager @ file:///${PROJECT_ROOT}/src/madsci_location_manager",
    "-e madsci.experiment_application @ file:///${PROJECT_ROOT}/src/madsci_experiment_application",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
    "*/venv/*",
    "*/.venv/*"
]
relative_files = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == '__main__':"
]
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
